requirements.txt
Django==4.2.7
langchain==0.2.1
langchain-google-genai==1.0.4
langchain-community==0.2.1
langchain-core==0.2.3
langchainhub
pymupdf==1.24.1
faiss-cpu==1.8.0
gunicorn
neo4j
sentence-transformers
google-generativeai
docling
django-q
python-dotenv
icecream
